requirements.txt
click==8.0.4
colorama==0.4.4
environs==9.5.0
Flask==2.0.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
marshmallow==3.14.1
psycopg2==2.9.3
PyJWT==1.7.1
python-dotenv==0.19.2
Werkzeug==2.0.3
